Implement staff logic unit tests

+++ src/main/java/org/movierental/staff/service/StaffService.java
@@ -17,7 +17,7 @@ public interface StaffService {
 
     List<Staff> findByLastname(String lastname);
 
-    List<Staff> findBySalaryRange(int min, int max);
+    List<Staff> findBySalaryRange(double min, double max);
 
     List<Staff> findByPositionId(Long positionId);
